WIP: Almost have everything switched to candid Int
The only problem is that Int lives on the heap, not the stack, so you
can't copy. So now I have to re-think everything...

+++ examples/motoko_examples/quicksort/canisters/rust/src/lib.rs
@@ -1,6 +1,7 @@
 mod quicksort;
+use ic_cdk::export::candid::Int;
 
 #[ic_cdk_macros::query]
-fn sort(xs: Vec<i128>) -> Vec<i128> {
-    quicksort::sort_by(xs, &i128::cmp)
+fn sort(xs: Vec<Int>) -> Vec<Int> {
+    quicksort::sort_by(xs, &Int::cmp)
 }
